Simplify abstraction for `sampling`
